# More efficient reading and writing of JSON documents