--- a/modules/json/test/src/smithy4s/http/json/SchematicJCodecTests.scala
+++ b/modules/json/test/src/smithy4s/http/json/SchematicJCodecTests.scala
@@ -475,7 +475,7 @@ object SchematicJCodecTests extends SimpleIOSuite {
       failure("Unexpected success")
     } catch {
       case PayloadError(_, _, message) =>
-        expect(message == "input Json document exceeded max arity of `1024`")
+        expect(message == "input JSON document exceeded max arity of `1024`")
     }
   }
 
@@ -487,7 +487,7 @@ object SchematicJCodecTests extends SimpleIOSuite {
       failure("Unexpected success")
     } catch {
       case PayloadError(_, _, message) =>
-        expect(message == "input Json document exceeded max arity of `1024`")
+        expect(message == "input JSON document exceeded max arity of `1024`")
     }
   }